fix(audit): pass 1 — comment corrections, NaN-safe clamp fallbacks, missed DRY sites, registry file split
Fix the inverted rotation-direction labels in common.glsl's rot() doc
comment and drop endeavouros-drift's stale sdSegment credit. Convert
the two byte-identical helpers the DRY pass missed (heat-melt's
classicHash copy, desktop-pixelate's inline crossFade) and grid's
remaining inline min-width read. Add the 'or <bound>' fallbacks that
pluau.clamp's contract requires in cascade, spread, and stair, matching
the old inline clamps' NaN results, and widen clamp's declared param
type to match its accepts-anything implementation. Rename opensuse's
forward-difference curl noise to curlNoiseFwd so it can never collide
with flow-noise.glsl, and replace liquid-canvas's shadowing local rot
matrix with the shared rot() helper. Move the builtin buffer resolver
into surfacebuiltinbuffers.cpp to keep surfaceshaderregistry.cpp under
the 800-line cap, canonicalize the QStandardPaths branch like the
sibling branch, fix the header doc's sibling-dir wording, and cover the
QStandardPaths fallback with a test.
